requirements.txt
Adafruit-Blinka==6.10.1
adafruit-circuitpython-fingerprint==2.2.0
Adafruit-PlatformDetect==3.13.3
Adafruit-PureIO==1.1.8
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
idna==2.10
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.3
opencv-python==4.5.1.48
Pillow==8.2.0
pkg-resources==0.0.0
pyftdi==0.53.1
pyserial==3.4
pyusb==1.1.1
PyYAML==5.4.1
requests==2.25.1
rpi==0.0.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
serial==0.0.97
six==1.16.0
sysv-ipc==1.1.0
urllib3==1.26.4
Werkzeug==1.0.1
